+++ src/backend_cairo/mod.rs
@@ -33,7 +33,6 @@ mod fill;
 mod filter;
 mod gradient;
 mod image;
-mod marker;
 mod mask;
 mod path;
 mod pattern;
@@ -83,7 +82,7 @@ impl ReCairoContextExt for cairo::Context {
             color.red as f64 / 255.0,
             color.green as f64 / 255.0,
             color.blue as f64 / 255.0,
-            *opacity,
+            opacity.value(),
         );
     }
 
@@ -261,7 +260,7 @@ fn render_node(
             Some(render_group(node, opt, layers, cr))
         }
         usvg::NodeKind::Path(ref path) => {
-            Some(path::draw(&node.tree(), path, opt, layers, cr))
+            Some(path::draw(&node.tree(), path, opt, cr))
         }
         usvg::NodeKind::Image(ref img) => {
             Some(image::draw(img, opt, cr))
@@ -352,8 +351,8 @@ fn render_group_impl(
     let curr_matrix = cr.get_matrix();
     cr.set_matrix(cairo::Matrix::identity());
     cr.set_source_surface(&*sub_surface, 0.0, 0.0);
-    if let Some(opacity) = g.opacity {
-        cr.paint_with_alpha(*opacity);
+    if !g.opacity.is_default() {
+        cr.paint_with_alpha(g.opacity.value());
     } else {
         cr.paint();
     }