+++ src/backend_qt/clippath.rs
@@ -39,7 +39,7 @@ pub fn apply(
 
         match *node.borrow() {
             usvg::NodeKind::Path(ref path_node) => {
-                super::path::draw(&node.tree(), path_node, opt, layers, &mut clip_p);
+                super::path::draw(&node.tree(), path_node, opt, &mut clip_p);
             }
             usvg::NodeKind::Group(ref g) => {
                 clip_group(&node, g, opt, bbox, layers, &mut clip_p);
@@ -85,7 +85,7 @@ fn clip_group(
 
                 let mut clip_p = qt::Painter::new(&mut clip_img);
                 clip_p.set_transform(&p.get_transform());
-                draw_group_child(&node, opt, layers, &mut clip_p);
+                draw_group_child(&node, opt, &mut clip_p);
 
                 apply(clip_node, cp, opt, bbox, layers, &mut clip_p);
                 clip_p.end();
@@ -101,14 +101,13 @@ fn clip_group(
 fn draw_group_child(
     node: &usvg::Node,
     opt: &Options,
-    layers: &mut QtLayers,
     p: &mut qt::Painter,
 ) {
     if let Some(child) = node.first_child() {
         p.apply_transform(&child.transform().to_native());
 
         if let usvg::NodeKind::Path(ref path_node) = *child.borrow() {
-            super::path::draw(&child.tree(), path_node, opt, layers, p);
+            super::path::draw(&child.tree(), path_node, opt, p);
         }
     }
 }