+++ usvg/src/lib.rs
@@ -152,3 +152,4 @@ pub use error::*;
 pub use geom::*;
 pub use options::*;
 pub use tree::*;
+pub use traits::IsDefault;