+++ usvg/src/tree/nodes.rs
@@ -20,7 +20,6 @@ pub enum NodeKind {
     ClipPath(ClipPath),
     Mask(Mask),
     Pattern(Pattern),
-    Marker(Marker),
     Filter(Filter),
     Path(Path),
     Image(Image),
@@ -41,7 +40,6 @@ impl NodeKind {
             NodeKind::ClipPath(ref e) => e.id.as_str(),
             NodeKind::Mask(ref e) => e.id.as_str(),
             NodeKind::Pattern(ref e) => e.id.as_str(),
-            NodeKind::Marker(ref e) => e.id.as_str(),
             NodeKind::Filter(ref e) => e.id.as_str(),
             NodeKind::Path(ref e) => e.id.as_str(),
             NodeKind::Image(ref e) => e.id.as_str(),
@@ -62,7 +60,6 @@ impl NodeKind {
             NodeKind::ClipPath(ref e) => e.transform,
             NodeKind::Mask(_) => Transform::default(),
             NodeKind::Pattern(ref e) => e.transform,
-            NodeKind::Marker(_) => Transform::default(),
             NodeKind::Filter(_) => Transform::default(),
             NodeKind::Path(ref e) => e.transform,
             NodeKind::Image(ref e) => e.transform,
@@ -112,9 +109,6 @@ pub struct Path {
 
     /// Stroke style.
     pub stroke: Option<Stroke>,
-
-    /// Marker.
-    pub marker: Box<PathMarker>,
 
     /// Segments list.
     ///
@@ -177,7 +171,7 @@ pub struct Group {
     ///
     /// After the group is rendered we should combine
     /// it with a parent group using the specified opacity.
-    pub opacity: Option<Opacity>,
+    pub opacity: Opacity,
 
     /// Element clip path.
     pub clip_path: Option<String>,
@@ -391,42 +385,6 @@ pub struct Pattern {
 
     /// Pattern viewbox.
     pub view_box: Option<ViewBox>,
-}
-
-
-/// A marker element.
-///
-/// `marker` element in SVG.
-#[derive(Clone, Debug)]
-pub struct Marker {
-    /// Element's ID.
-    ///
-    /// Taken from the SVG itself.
-    /// Can't be empty.
-    pub id: String,
-
-    /// Coordinate system units.
-    ///
-    /// `markerUnits` in SVG.
-    pub units: MarkerUnits,
-
-    /// Marker rectangle.
-    ///
-    /// `refX`, `refY`, `markerWidth` and `markerHeight` in SVG.
-    pub rect: Rect,
-
-    /// Marker viewbox.
-    pub view_box: Option<ViewBox>,
-
-    /// Marker orientation.
-    ///
-    /// `orient` in SVG.
-    pub orientation: MarkerOrientation,
-
-    /// Marker overflow.
-    ///
-    /// `overflow` in SVG.
-    pub overflow: Overflow,
 }